// demo-mcp-21-code-generator-node/input/reference-repos/MSServicehomepass/src/common/utils/enums/tracing.enum.ts
/**
 * Enumeraciones usadas internamente para describir distintas tareas dentro del flujo del MS
 * @author Sebastian Traslaviña
 */
export enum ETaskTracing {
  RECEIVED = 'RECEIVED',
  PAYMENT_RECEIVED = 'PAYMENT RECEIVED',
  RECEIVED_FAILED = 'RESERVATION FAILED',
  INVALID_CHANNEL = 'INVALID CHANNEL',
  INVALID_FORMAT = 'INVALID FORMAT',
  ORDER_EXISTS = 'ORDER ALREADY EXISTS',
  ORDER_NOT_EXIST = 'ORDER DOESNT EXISTS',
  SP_NOT_EXIST = 'SALES PRODUCT DOESNT EXIST',
  FLOW_FAILED = 'GET FLOW FAILED',
  PAYMENT_SUCCESS = 'PAYMENT CONFIRMATION SUCCESFULL',
  PAYMENT_FAILED = 'PAYMENT CONFIRMATION FAILED',
  MISSING_PRODUCT = 'MISSING REQUIRED INFO PRODUCT',
  INVALID_FLAG = 'INVALID FLAG IN PARAM',
  INVALID_FIELDS_INSPIRA = 'INVALID REQUIRED FIELDS TO INSPIRA',
  INVALID_FIELDS_OPERATION = 'INVALID REQUIRED FIELDS TO OPERATION',
  INVALID_REQUIRED = 'INVALID REQUIRED FIELDS',
  ORDER_SAVED = 'STORED',
  SHOPPING_CART_ACCEPT_SUCCESS = 'SHOPPING CART ACCEPT SUCCESSFULL',
  SHOPPING_CART_ACCEPT_FAILED = 'SHOPPING CART ACCEPT FAILED',
  SHOPPING_CART_CONFIRM_SUCCESS = 'SHOPPING CART CONFIRM SUCCESSFULL',
  SHOPPING_CART_CONFIRM_FAILED = 'SHOPPING CART CONFIRM FAILED',
  SHOPPING_CART_UPDATE_ITEM_SUCCESS = 'SHOPPING CART UPDATE ITEM SUCCESSFULL',
  SHOPPING_CART_UPDATE_ITEM_FAILED = 'SHOPPING CART UPDATE ITEM FAILED',
  NUMBER_RESERVE_FAILED = 'NUMBER RESERVE FAILED',
  CONFIRM_NUMBER_RESERVE_FAILED = 'CONFIRM NUMBER RESERVE FAILED',
  INVALID_TYPEPRODUCT = 'INVALID TYPEPRODUCT',
  MS_MATRIZ = 'MS MATRIZ',
  MS_HOMOLOGATES = 'MS HOMOLOGATES',
  SP_CODE_VALIDATE = 'VALIDATE SP CODE',
}

export enum EStatusTracingGeneral {
  STATUS_PENDIENTE = 'PENDIENTE',
  FAILED = 'FAILED',
  ERROR = 'ERROR',
  STATUS_SUCCESS = 'SUCCESS',
  STATUS_FAILED = 'FAILED',
  LEGACY_SUCCESS = 'LEGACY_SUCCESS',
  LEGACY_ERROR = 'LEGACY_ERROR',
  LEGACY_WARN = 'LEGACY_WARN',
  BD_SUCCESS = 'BD_SUCCESS',
  BD_ERROR = 'BD_ERROR',
  BD_WARN = 'BD_WARN',
}

export enum EDescriptionTracing {
  ERROR_ENDPOINT_GLOBAL_DES = 'Error servicio de EndPoint',
  ERROR_RESPONSE_GLOBAL_DES = 'Error respuesta de request',
  INVALID_CHANNEL_GLOBAL_DES = 'Se valida canal del request recibido',
  INVALID_FORMAT_HTTP_DES = 'Se valida formato y campos del request recibido',
  MS_HOMOLOGATES_DES = 'Consumo servicio Ms-prod-conf-homologates',
  MS_HOMOLOGATES_VALIDATE_TMCODE_DES = 'Validate tm_code servicio Ms-prod-conf-homologates',
  MS_MATRIZ_DES = 'Consumo servicio Ms-matriz',
  SP_CODE_VALIDATE_DES = 'Validación SpCode existe en información ms-matriz',
  VALIDATE_TYPE_PRODUCT = 'Se valida typeProduct del request recibido',
}

export enum ETaskTracingGeneral {
  CONSUMIR_CMATRICES = 'CONSULTAR LEGADO CMATRICESAS400',
  DELETE_DATA = 'ELIMINACION DE DATOS',
  DELETE_DATA_MOVIL = 'ELIMINACION DE DATOS',
  DOWNLOAD_DATA = 'DESCARGA DE DATOS',
  EXEC_CRON = 'EJECUCION CRON',
  EXEC_MANUAL = 'EJECUCION MANUAL',
  FILE_DOWNLOAD = 'DESCARGANDO ARCHIVO',
  FINAL_REQUEST = 'FINAL_REQUEST',
  GET_CUSREQUESTHOMEPASS = 'CONSULTAR DATOS EN CUSREQUESTHOMEPASS',
  INICIO_REQUEST = 'INICIO_REQUEST',
  INVALID_CHANNEL = 'INVALID CHANNEL',
  SAVE_DATA = 'GUARDADO DE DATOS',
  SAVE_DATA_MOVIL = 'GUARDADO DE DATOS',
  SOLICITUD_POR_ID = 'BUSCAR SOLICITUD POR ID SOLICITUD',
  TASK_GENERAL = 'TEST',
  TRACING_CAPACITY = 'CONSUMO DEL END POINT CAPACITY',
  TRACING_CREATE_ORDER = 'CONSUMO DEL END POINT CREATE ORDER',
  TRACING_PAYMENT = 'CONSUMO DEL END POINT PAYMENT',
  TRACING_SENDORDER = 'CONSUMO DEL END POINT SEND ORDER',
  UPDATE_CUSREQUESTHOMEPASS = 'ACTUALIZAR DATOS EN CUSREQUESTHOMEPASS',
}